third_party/WebKit/LayoutTests/device_orientation/motion/page-visibility.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/device-orientation-helpers.js"></script>
<script>
'use strict';

async_test(test => {
  assertTestRunner();

  var motionData = generateMotionData(0, 0, 0,
                                      0, 0, 0,
                                      0, 0, 0,
                                      0);
  function succeedAndFinish()
  {
    test.done();
  }

  function testWithPageVisible()
  {
    window.removeEventListener('devicemotion', failAndFinish);
    testRunner.setPageVisibility("visible");
    window.addEventListener('devicemotion', succeedAndFinish);
  }

  function failAndFinish()
  {
    assert_true(false);
    test.done();
  }

  function testWithPageHidden()
  {
    window.removeEventListener('devicemotion', deviceMotionListener);
    testRunner.setPageVisibility("hidden");
    window.addEventListener('devicemotion', failAndFinish);
    setTimeout(testWithPageVisible, 100);
  }

  function deviceMotionListener(event) {
    setTimeout(testWithPageHidden, 0);
  }

  setMockMotion(motionData);
  window.addEventListener('devicemotion', deviceMotionListener);
}, 'Tests to check that devicemotion events are not fired when the page is not visible.');
</script>
</body>
</html>
